Add keibabutton
