+++ src/pages/index.tsx
@@ -1,13 +1,5 @@
-import styled from '@emotion/styled';
+import Top from 'components/Organisms/Top';
 
-const Contents = styled.div`
-  background: #151515 url('/images/login.jpg') no-repeat scroll center bottom;
-  background-size: cover;
-  color: #fff;
-  text-align: center;
-  height: 90vh;
-`;
-
-export default function Top() {
-  return <Contents />;
+export default function TopPage() {
+  return <Top />;
 }
